Update the Django example
