+++ examples/django/manage.py
@@ -5,7 +5,8 @@ import sys
 
 
 def main():
-    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "slackapp.settings")
+    """Run administrative tasks."""
+    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "myslackapp.settings")
     try:
         from django.core.management import execute_from_command_line
     except ImportError as exc:
