+++ examples/django/oauth_app/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class OauthAppConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "oauth_app"
