+++ examples/django/oauth_app/slack_datastores.py
@@ -1,78 +1,6 @@
 # ----------------------
-# Database tables
-# ----------------------
-
-from django.db import models
-
-
-class SlackBot(models.Model):
-    client_id = models.CharField(null=False, max_length=32)
-    app_id = models.CharField(null=False, max_length=32)
-    enterprise_id = models.CharField(null=True, max_length=32)
-    enterprise_name = models.TextField(null=True)
-    team_id = models.CharField(null=True, max_length=32)
-    team_name = models.TextField(null=True)
-    bot_token = models.TextField(null=True)
-    bot_id = models.CharField(null=True, max_length=32)
-    bot_user_id = models.CharField(null=True, max_length=32)
-    bot_scopes = models.TextField(null=True)
-    is_enterprise_install = models.BooleanField(null=True)
-    installed_at = models.DateTimeField(null=False)
-
-    class Meta:
-        indexes = [
-            models.Index(
-                fields=["client_id", "enterprise_id", "team_id", "installed_at"]
-            ),
-        ]
-
-
-class SlackInstallation(models.Model):
-    client_id = models.CharField(null=False, max_length=32)
-    app_id = models.CharField(null=False, max_length=32)
-    enterprise_id = models.CharField(null=True, max_length=32)
-    enterprise_name = models.TextField(null=True)
-    enterprise_url = models.TextField(null=True)
-    team_id = models.CharField(null=True, max_length=32)
-    team_name = models.TextField(null=True)
-    bot_token = models.TextField(null=True)
-    bot_id = models.CharField(null=True, max_length=32)
-    bot_user_id = models.TextField(null=True)
-    bot_scopes = models.TextField(null=True)
-    user_id = models.CharField(null=False, max_length=32)
-    user_token = models.TextField(null=True)
-    user_scopes = models.TextField(null=True)
-    incoming_webhook_url = models.TextField(null=True)
-    incoming_webhook_channel = models.TextField(null=True)
-    incoming_webhook_channel_id = models.TextField(null=True)
-    incoming_webhook_configuration_url = models.TextField(null=True)
-    is_enterprise_install = models.BooleanField(null=True)
-    token_type = models.CharField(null=True, max_length=32)
-    installed_at = models.DateTimeField(null=False)
-
-    class Meta:
-        indexes = [
-            models.Index(
-                fields=[
-                    "client_id",
-                    "enterprise_id",
-                    "team_id",
-                    "user_id",
-                    "installed_at",
-                ]
-            ),
-        ]
-
-
-class SlackOAuthState(models.Model):
-    state = models.CharField(null=False, max_length=64)
-    expire_at = models.DateTimeField(null=False)
-
-
-# ----------------------
 # Bolt store implementations
 # ----------------------
-
 
 from logging import Logger
 from typing import Optional
@@ -82,7 +10,8 @@ from django.utils import timezone
 from django.utils.timezone import is_naive, make_aware
 from slack_sdk.oauth import InstallationStore, OAuthStateStore
 from slack_sdk.oauth.installation_store import Bot, Installation
-from slack_sdk.webhook import WebhookClient
+
+from .models import SlackBot, SlackInstallation, SlackOAuthState
 
 
 class DjangoInstallationStore(InstallationStore):
@@ -220,71 +149,3 @@ class DjangoOAuthStateStore(OAuthStateStore):
                 row.delete()
             return True
         return False
-
-
-# ----------------------
-# Slack App
-# ----------------------
-
-import logging
-import os
-from slack_bolt import App, BoltContext
-from slack_bolt.oauth.oauth_settings import OAuthSettings
-
-logger = logging.getLogger(__name__)
-client_id, client_secret, signing_secret, scopes = (
-    os.environ["SLACK_CLIENT_ID"],
-    os.environ["SLACK_CLIENT_SECRET"],
-    os.environ["SLACK_SIGNING_SECRET"],
-    os.environ.get("SLACK_SCOPES", "commands").split(","),
-)
-
-app = App(
-    signing_secret=signing_secret,
-    installation_store=DjangoInstallationStore(
-        client_id=client_id,
-        logger=logger,
-    ),
-    oauth_settings=OAuthSettings(
-        client_id=client_id,
-        client_secret=client_secret,
-        scopes=scopes,
-        state_store=DjangoOAuthStateStore(
-            expiration_seconds=120,
-            logger=logger,
-        ),
-    ),
-)
-
-
-def event_test(body, say, context: BoltContext, logger):
-    logger.info(body)
-    say(":wave: What's up?")
-
-    found_rows = list(
-        SlackInstallation.objects.filter(enterprise_id=context.enterprise_id)
-        .filter(team_id=context.team_id)
-        .filter(incoming_webhook_url__isnull=False)
-        .order_by(F("installed_at").desc())[:1]
-    )
-    if len(found_rows) > 0:
-        webhook_url = found_rows[0].incoming_webhook_url
-        logger.info(f"webhook_url: {webhook_url}")
-        client = WebhookClient(webhook_url)
-        client.send(text=":wave: This is a message posted using Incoming Webhook!")
-
-
-# lazy listener example
-def noop():
-    pass
-
-
-app.event("app_mention")(
-    ack=event_test,
-    lazy=[noop],
-)
-
-
-@app.command("/hello-django-app")
-def command(ack):
-    ack(":wave: Hello from a Django app :smile:")
