+++ examples/django/simple_app/slack_listeners.py
@@ -1,0 +1,19 @@
+import logging
+import os
+
+from slack_bolt import App
+
+logger = logging.getLogger(__name__)
+
+app = App(
+    token=os.environ["SLACK_BOT_TOKEN"],
+    signing_secret=os.environ["SLACK_SIGNING_SECRET"],
+    # disable eagerly verifying the given SLACK_BOT_TOKEN value
+    token_verification_enabled=False,
+)
+
+
+@app.event("app_mention")
+def handle_app_mentions(logger, event, say):
+    logger.info(event)
+    say(f"Hi there, <@{event['user']}>")
